move fixtures to normal file
